+++ src/driver/sqlx_postgres.rs
@@ -602,7 +602,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<ipnetwork::IpNetwork>, _>(c.ordinal())
                                     .expect("Failed to get ip address array")
                                     .iter()
-                                    .map(|val| Value::IpNetwork(Some(Box::new(val.clone()))))
+                                    .map(|val| Value::IpNetwork(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),
@@ -639,7 +639,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<chrono::NaiveDateTime>, _>(c.ordinal())
                                     .expect("Failed to get timestamp array")
                                     .iter()
-                                    .map(|val| Value::ChronoDateTime(Some(Box::new(val.clone()))))
+                                    .map(|val| Value::ChronoDateTime(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),
@@ -675,7 +675,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<chrono::NaiveDate>, _>(c.ordinal())
                                     .expect("Failed to get date array")
                                     .iter()
-                                    .map(|val| Value::ChronoDate(Some(Box::new(val.clone()))))
+                                    .map(|val| Value::ChronoDate(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),
@@ -711,7 +711,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<chrono::NaiveTime>, _>(c.ordinal())
                                     .expect("Failed to get time array")
                                     .iter()
-                                    .map(|val| Value::ChronoTime(Some(Box::new(val.clone()))))
+                                    .map(|val| Value::ChronoTime(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),
@@ -747,9 +747,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<chrono::DateTime<chrono::Utc>>, _>(c.ordinal())
                                     .expect("Failed to get timestamptz array")
                                     .iter()
-                                    .map(|val| {
-                                        Value::ChronoDateTimeUtc(Some(Box::new(val.clone())))
-                                    })
+                                    .map(|val| Value::ChronoDateTimeUtc(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),
@@ -785,7 +783,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<chrono::NaiveTime>, _>(c.ordinal())
                                     .expect("Failed to get timetz array")
                                     .iter()
-                                    .map(|val| Value::ChronoTime(Some(Box::new(val.clone()))))
+                                    .map(|val| Value::ChronoTime(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),
@@ -817,7 +815,7 @@ pub(crate) fn from_sqlx_postgres_row_to_proxy_row(row: &sqlx::postgres::PgRow) -
                                 row.try_get::<Vec<uuid::Uuid>, _>(c.ordinal())
                                     .expect("Failed to get uuid array")
                                     .iter()
-                                    .map(|val| Value::Uuid(Some(Box::new(val.clone()))))
+                                    .map(|val| Value::Uuid(Some(Box::new(*val))))
                                     .collect(),
                             )),
                         ),