+++ CHANGELOG.md
@@ -5,6 +5,99 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.16 - 2025-09-11
+
+### Bug Fixes
+
+* Fix enum casting in DerivePartialModel https://github.com/SeaQL/sea-orm/pull/2719 https://github.com/SeaQL/sea-orm/pull/2720
+```rust
+#[derive(DerivePartialModel)]
+#[sea_orm(entity = "active_enum::Entity", from_query_result, alias = "zzz")]
+struct PartialWithEnumAndAlias {
+    #[sea_orm(from_col = "tea")]
+    foo: Option<Tea>,
+}
+
+let sql = active_enum::Entity::find()
+    .into_partial_model::<PartialWithEnumAndAlias>()
+    .into_statement(DbBackend::Postgres)
+    .sql;
+
+assert_eq!(
+    sql,
+    r#"SELECT CAST("zzz"."tea" AS "text") AS "foo" FROM "public"."active_enum""#,
+);
+```
+
+### Enhancements
+
+* [sea-orm-cli] Use tokio (optional) instead of async-std https://github.com/SeaQL/sea-orm/pull/2721
+
+## 1.1.15 - 2025-08-31
+
+### Enhancements
+
+* Allow `DerivePartialModel` to have nested aliases https://github.com/SeaQL/sea-orm/pull/2686
+```rust
+#[derive(DerivePartialModel)]
+#[sea_orm(entity = "bakery::Entity", from_query_result)]
+struct Factory {
+    id: i32,
+    #[sea_orm(from_col = "name")]
+    plant: String,
+}
+
+#[derive(DerivePartialModel)]
+#[sea_orm(entity = "cake::Entity", from_query_result)]
+struct CakeFactory {
+    id: i32,
+    name: String,
+    #[sea_orm(nested, alias = "factory")] // <- new
+    bakery: Option<Factory>,
+}
+```
+* Add `ActiveModelTrait::try_set` https://github.com/SeaQL/sea-orm/pull/2706
+```rust
+fn set(&mut self, c: <Self::Entity as EntityTrait>::Column, v: Value);
+/// New: a non-panicking version of above
+fn try_set(&mut self, c: <Self::Entity as EntityTrait>::Column, v: Value) -> Result<(), DbErr>;
+```
+
+### Bug Fixes
+
+* [sea-orm-cli] Fix compilation issue https://github.com/SeaQL/sea-orm/pull/2713
+
+## 1.1.14 - 2025-07-21
+
+### Enhancements
+
+* [sea-orm-cli] Mask sensitive ENV values https://github.com/SeaQL/sea-orm/pull/2658
+
+### Bug Fixes
+
+* `FromJsonQueryResult`: panic on serialization failures https://github.com/SeaQL/sea-orm/pull/2635
+```rust
+#[derive(Clone, Debug, PartialEq, Deserialize, FromJsonQueryResult)]
+pub struct NonSerializableStruct;
+
+impl Serialize for NonSerializableStruct {
+    fn serialize<S>(&self, _serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        Err(serde::ser::Error::custom(
+            "intentionally failing serialization",
+        ))
+    }
+}
+
+let model = Model {
+    json: Some(NonSerializableStruct),
+};
+
+let _ = model.into_active_model().insert(&ctx.db).await; // panic here
+```
+
 ## 1.1.13 - 2025-06-29
 
 ### New Features